requirements.txt
streamlit==1.28.0
tensorflow==2.20.0rc0
Pillow==10.0.1
numpy==1.26.0
protobuf==4.23.4
